Set modes for the two etc entries in the installer
While `/etc/nixos-variables/system.nix` ends up as a symlink to
`/etc/static/nixos-variables/system.nix` and could go either way, `/etc/nasty`
ends up as a symlink to `/etc/staticy/nasty` which itself symlinks to the now
non-existent GitHub workspace.

Since `/etc/nixos` itself is a real directory (not a symlink), we should be able
to solve that and make these directories match that style.