Preserve SSH host keys across reinstalls
Secrets are going to be encrypted using the host SSH keys so those need to
be loaded onto the system somehow. Possible options I could come up with:

1. Load them from a second USB key inserted along with the installer image
2. Load them from `/data`, but would require importing the ZFS pool
3. Load them from a new, long-term partition on the root drive

USB key will work for now. Also means it's easy to recopy from my password
store if the key dies or I otherwise need to replace it.