Unset the token leftover from the GitHub checkout action
GitHub runners don't have the git CLI preinstalled but also there doesn't seem
to be wildcard support in the `--unset-all` or `--remove-section` commands. I
don't want to hardcode the section name in case there are times when it might
or might not have a trailing slash in the `http.https://github.com/` config
section.