Block remote access to the health and metrics Jellyfin endpoints
This does result in those endpoints returning a 404 but maybe that's not
entirely incorrect since, as far as the outside world is concerned, those
endpoints don't exist. Or is that me justifying not writing a few extra lines
right now to make it a 403?

In the course of figuring this out, I found out you can set the cert to serve
globally at the entrypoint level, so does that too.

Still need to figure out how to enable the metrics from NixOS configs.