Enable Jellyfin metrics
I worked off https://github.com/matt1432/nixos-jellyfin as a reference, but
doing all the simplest things possible. I'm sure there are issues here, hence
the complexity in that repo, but it's a start and this is how I'll learn.

Only do system.xml for now, all I want immediately is metrics.